Pulse-Ops/pyproject.toml
[tool.poetry]
name = "pulse-ops"
version = "0.1.0"
description = "Pulse-Ops Backend and Agents Monorepo"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "backend" },
    { include = "agents" },
    { include = "shared" },
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
pydantic = "^2.4.2"
pydantic-ai = "^0.0.1" # Placeholder for PydanticAI
python-dotenv = "^1.0.0"
supabase = "^2.0.2"
httpx = "^0.27.0"
slowapi = "^0.1.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

packages = [
    { include = "backend" },
    { include = "agents" },
    { include = "shared" },
]
